codebase/requirements.txt
plotly
flask
flask_oauth
httplib2
apiclient
oauth2client
email
dill
gunicorn
spacy==1.9.0
nose
future
importlib
argparse
python-dateutil
pytz
gevent
vaderSentiment
sqlalchemy
Flask-SQLAlchemy
Flask-Migrate
psycopg2
pandas==0.20.3
langdetect
numpy==1.9.0
nltk==3.0.0
scipy==0.19.1
scikit-learn==0.15.1
pycorenlp==0.3.0
google-api-python-client 
Flask-HTTPAuth